+++ app/src/main/java/ru/mugalimov/volthome/model/Device.kt
@@ -3,11 +3,11 @@ package ru.mugalimov.volthome.model
 import java.util.Date
 
 data class Device(
-    val id: Int,
+    val id: Long,
     val name: String,
     val power: Int, //мощность
     val voltage: Int, //Класс напряжения
     val demandRatio: Double, //к-т спроса
-    val roomId: Int,
+    val roomId: Long,
     val createdAt: Date
 )
