+++ app/src/main/java/ru/mugalimov/volthome/repository/DeviceRepository.kt
@@ -6,13 +6,13 @@ import ru.mugalimov.volthome.model.Device
 
 interface DeviceRepository {
     //получаем список устройств по id комнаты
-    suspend fun observeDevicesByIdRoom(roomId: Int) : Flow<List<Device>>
+    suspend fun observeDevicesByIdRoom(roomId: Long) : Flow<List<Device>>
 
     //добавить устройство в комнату
-    suspend fun addDevice(name: String, power: Int, voltage: Int, demandRatio: Double, roomId: Int)
+    suspend fun addDevice(name: String, power: Int, voltage: Int, demandRatio: Double, roomId: Long)
 
     //удалить комнату
-    suspend fun deleteDevice(deviceId: Int)
+    suspend fun deleteDevice(deviceId: Long)
 
     //получить комнату по roomId
     suspend fun getDeviceById(deviceId: Int) : Device?
